Refactor Semantic Scholar client
